finish dashboard refactor
